transcription-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
kafka-python-ng==2.2.2
minio==7.2.5
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
openai-whisper==20231117
numpy==1.26.4
prometheus-client==0.20.0
